Put auth response mapping stuff into own file
